Refactored Compute to Computation. Made it non-static.
