+++ UnitTests/TestDPS.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using POESKillTree.SkillTreeFiles;
+using POESKillTree.Compute;
 
 namespace UnitTests
 {
@@ -56,13 +57,13 @@ namespace UnitTests
                     return value;
 
                 case Rounding.Round:
-                    return Compute.RoundValue(value, variant.Precision);
+                    return ComputeGlobal.RoundValue(value, variant.Precision);
 
                 case Rounding.Floor:
-                    return Compute.FloorValue(value, variant.Precision);
+                    return ComputeGlobal.FloorValue(value, variant.Precision);
 
                 case Rounding.Ceil:
-                    return Compute.CeilValue(value, variant.Precision);
+                    return ComputeGlobal.CeilValue(value, variant.Precision);
             }
 
             throw new NotImplementedException();
@@ -166,16 +167,15 @@ namespace UnitTests
         public void TestRoundingHalfDown()
         {
             // Round Half Toward Zero (Round Half Down).
+            Assert.AreEqual(1.36f, ComputeGlobal.RoundHalfDownEvenValue(1.365f, 2));
+            Assert.AreEqual(1.38f, ComputeGlobal.RoundHalfDownEvenValue(1.375f, 2));
 
-            Assert.AreEqual(1.36f, Compute.RoundHalfDownEvenValue(1.365f, 2));
-            Assert.AreEqual(1.38f, Compute.RoundHalfDownEvenValue(1.375f, 2));
-
-            Assert.AreEqual(27.9f, Compute.RoundHalfDownValue(27.94f, 1));
-            Assert.AreEqual(-27.9f, Compute.RoundHalfDownValue(-27.95f, 1));
-            Assert.AreEqual(27.9f, Compute.RoundHalfDownValue(27.95f, 1));
-            Assert.AreEqual(28f, Compute.RoundHalfDownValue(27.951f, 1));
-            Assert.AreEqual(28f, Compute.RoundHalfDownValue(27.96f, 1));
-            Assert.AreEqual(5f, Compute.RoundHalfDownValue(4.96f, 1));
+            Assert.AreEqual(27.9f, ComputeGlobal.RoundHalfDownValue(27.94f, 1));
+            Assert.AreEqual(-27.9f, ComputeGlobal.RoundHalfDownValue(-27.95f, 1));
+            Assert.AreEqual(27.9f, ComputeGlobal.RoundHalfDownValue(27.95f, 1));
+            Assert.AreEqual(28f, ComputeGlobal.RoundHalfDownValue(27.951f, 1));
+            Assert.AreEqual(28f, ComputeGlobal.RoundHalfDownValue(27.96f, 1));
+            Assert.AreEqual(5f, ComputeGlobal.RoundHalfDownValue(4.96f, 1));
         }
     }
 }
